Create README.md
IJSA mission: international cooperation of scientists conducting research in the field of integration of human sciences and health. IJSA publishes research examining the Human as an exposure as well as an outcome. As an exposure, the Journal publishes articles examining how socium influences all aspects of human life and its health. As an outcome, the Journal invites articles that examine the pedagogical, psychological, medical interventions that may affect human development and its health. IJSA is a peer-reviewed and publishes Review Articles and Original Research in fields of Social Sciences, Psychology and Health Sciences.